fix: unable to resolve imports for web fixes #58
