+++ .circleci/config.yml
@@ -24,10 +24,6 @@ jobs:
             - dependencies-
       - restore_cache:
           keys:
-            - dependencies-showcase-{{ checksum "showcase/package.json" }}
-            - dependencies-showcase-
-      - restore_cache:
-          keys:
             - dependencies-example-{{ checksum "example/package.json" }}
             - dependencies-example-
       - run:
@@ -39,9 +35,6 @@ jobs:
       - save_cache:
           key: dependencies-{{ checksum "package.json" }}
           paths: node_modules
-      - save_cache:
-          key: dependencies-showcase-{{ checksum "showcase/package.json" }}
-          paths: showcase/node_modules
       - save_cache:
           key: dependencies-example-{{ checksum "example/package.json" }}
           paths: example/node_modules
