+++ .npmignore
@@ -1,2 +1,3 @@
 /src
 lib/**/*.map
+/example
