+++ example/package.json
@@ -1,27 +1,31 @@
 {
   "name": "phosphor-react-native-example",
   "version": "1.0.0",
-  "main": "expo/AppEntry.js",
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
-    "bundle-vis": "EXPO_UNSTABLE_ATLAS=true npx expo start --no-dev"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "expo": "~51.0.26",
+    "expo": "^51.0.26",
     "expo-status-bar": "~1.12.1",
-    "phosphor-react-native": "^2.1.0",
+    "phosphor-react-native": "file:..",
     "react": "18.2.0",
     "react-native": "0.74.5",
-    "react-native-svg": "^15.5.0"
+    "react-native-svg": "15.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@types/react": "~18.2.45",
-    "expo-atlas": "^0.3.0",
-    "typescript": "^5.1.3"
+    "@babel/core": "^7.24.0",
+    "@eslint/js": "^9.9.0",
+    "@types/react": "~18.2.79",
+    "eslint": "9.x",
+    "eslint-plugin-react": "^7.35.0",
+    "globals": "^15.9.0",
+    "typescript": "~5.3.3",
+    "typescript-eslint": "^8.0.1"
   },
   "private": true
 }
