+++ example/phosphor-mark-tight-yellow.png
