+++ package.json
@@ -2,8 +2,8 @@
   "name": "phosphor-react-native",
   "version": "2.0.0",
   "description": "Flexible icons for React Native",
-  "main": "lib/commonjs/lib/index.js",
-  "module": "lib/module/lib/index.js",
+  "main": "lib/commonjs/index.js",
+  "module": "lib/module/index.js",
   "types": "lib/typescript/index.d.ts",
   "react-native": "src/index.tsx",
   "source": "src/index",
