--- showcase/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "phosphor-react-native-showcase",
-  "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "lint": "eslint ."
-  },
-  "dependencies": {
-    "expo": "^51.0.26",
-    "expo-status-bar": "~1.12.1",
-    "phosphor-react-native": "file:..",
-    "react": "18.2.0",
-    "react-native": "0.74.5",
-    "react-native-svg": "15.2.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@eslint/js": "^9.9.0",
-    "@types/react": "~18.2.79",
-    "eslint": "9.x",
-    "eslint-plugin-react": "^7.35.0",
-    "globals": "^15.9.0",
-    "typescript": "~5.3.3",
-    "typescript-eslint": "^8.0.1"
-  },
-  "private": true
-}
